+++ rake.d/target.rb
@@ -162,7 +162,6 @@ class Target
       when :intl             then c(:LIBINTL_LIBS)
       when :cmark            then c(:CMARK_LIBS)
       when :dvdread          then c(:DVDREAD_LIBS)
-      when :pcre2            then c(:PCRE2_LIBS)
       when :pugixml          then c?(:PUGIXML_INTERNAL) ? [ '-Llib/pugixml/src', '-lpugixml' ] : c(:PUGIXML_LIBS)
       when :qt               then c(:QT_LIBS)
       when :qt_non_gui       then c(:QT_LIBS_NON_GUI)